refactor: move App component into its own file and then importing it in index.js
